Updates for Leopard, including fast enumeration and NSInteger/NSUInteger. Added abstract parent classes for Stack and Queue implementations.
git-svn-id: http://dysart.cs.byu.edu/chsvn/trunk@17 a75b6edf-6070-4da8-89c8-9992cc802e51

+++ source/AbstractTree.h
@@ -34,7 +34,7 @@
 @interface AbstractTree : NSObject <Tree>
 {
 	/** A count of how many elements are currently in the tree. */
-	unsigned int count;
+	NSUInteger count;
 }
 
 - (NSEnumerator *)objectEnumerator;
@@ -57,7 +57,7 @@
 #pragma mark Inherited Methods
 
 - (void) addObjectsFromArray:(NSArray *)anArray;
-- (unsigned int) count;
+- (NSUInteger) count;
 
 - (NSSet *) contentsAsSet;
 - (NSArray *) contentsAsArrayWithOrder:(CHTraversalOrder)traversalOrder;
